requirements.txt
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.6
gunicorn
python-dotenv
djangorestframework-simplejwt
requests
django-cors-headers
drf-yasg
pytest